pyproject.toml
[tool.poetry]
name = "touring-api"
version = "0.1.0"
description = "Touring App REST API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-health-check = "^3.18.3"
django-extensions = "^3.2.3"
python-decouple = "^3.8"
django-auditlog = "^3.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
rav = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
